refactor: 컨벤션에 맞게 스타일링 분리
